update from team repo

+++ src/main/java/seedu/address/logic/commands/CheckExpenditureCommand.java
@@ -77,48 +77,48 @@ public class CheckExpenditureCommand extends Command {
             if ((year1 < year) && (year2 > year)) {
                 total = total + Float.parseFloat(editedExpenditure.getMoney().toString());
                 dpd.setValue(editedExpenditure.getDescription().toString() + ":"
-                                + Integer.parseInt(editedExpenditure.getMoney().toString()),
-                        Integer.parseInt(editedExpenditure.getMoney().toString()));
+                                + Double.parseDouble(editedExpenditure.getMoney().toString()),
+                        Double.parseDouble(editedExpenditure.getMoney().toString()));
             } else if ((year1 == year) && (year2 == year)) {
                 if ((month1 < month) && (month2 > month)) {
                     total = total + Float.parseFloat(editedExpenditure.getMoney().toString());
                     dpd.setValue(editedExpenditure.getDescription().toString() + ":"
-                                    + Integer.parseInt(editedExpenditure.getMoney().toString()),
-                            Integer.parseInt(editedExpenditure.getMoney().toString()));
+                                    + Double.parseDouble(editedExpenditure.getMoney().toString()),
+                            Double.parseDouble(editedExpenditure.getMoney().toString()));
                 } else if ((month1 == month) && (day1 <= day)) {
                     total = total + Float.parseFloat(editedExpenditure.getMoney().toString());
                     dpd.setValue(editedExpenditure.getDescription().toString() + ":"
-                                    + Integer.parseInt(editedExpenditure.getMoney().toString()),
-                            Integer.parseInt(editedExpenditure.getMoney().toString()));
+                                    + Double.parseDouble(editedExpenditure.getMoney().toString()),
+                            Double.parseDouble(editedExpenditure.getMoney().toString()));
                 } else if ((month2 == month) && (day2 >= day)) {
                     total = total + Float.parseFloat(editedExpenditure.getMoney().toString());
                     dpd.setValue(editedExpenditure.getDescription().toString() + ":"
-                                    + Integer.parseInt(editedExpenditure.getMoney().toString()),
-                            Integer.parseInt(editedExpenditure.getMoney().toString()));
+                                    + Double.parseDouble(editedExpenditure.getMoney().toString()),
+                            Double.parseDouble(editedExpenditure.getMoney().toString()));
                 }
             } else if ((year1 == year) && (year2 > year)) {
                 if (month1 < month) {
                     total = total + Float.parseFloat(editedExpenditure.getMoney().toString());
                     dpd.setValue(editedExpenditure.getDescription().toString() + ":"
-                                    + Integer.parseInt(editedExpenditure.getMoney().toString()),
-                            Integer.parseInt(editedExpenditure.getMoney().toString()));
+                                    + Double.parseDouble(editedExpenditure.getMoney().toString()),
+                            Double.parseDouble(editedExpenditure.getMoney().toString()));
                 } else if ((month1 == month) && (day1 <= day)) {
                     total = total + Float.parseFloat(editedExpenditure.getMoney().toString());
                     dpd.setValue(editedExpenditure.getDescription().toString() + ":"
-                                    + Integer.parseInt(editedExpenditure.getMoney().toString()),
-                            Integer.parseInt(editedExpenditure.getMoney().toString()));
+                                    + Double.parseDouble(editedExpenditure.getMoney().toString()),
+                            Double.parseDouble(editedExpenditure.getMoney().toString()));
                 }
             } else if ((year1 < year) && (year2 == year)) {
                 if (month2 > month) {
                     total = total + Float.parseFloat(editedExpenditure.getMoney().toString());
                     dpd.setValue(editedExpenditure.getDescription().toString() + ":"
-                                    + Integer.parseInt(editedExpenditure.getMoney().toString()),
-                            Integer.parseInt(editedExpenditure.getMoney().toString()));
+                                    + Double.parseDouble(editedExpenditure.getMoney().toString()),
+                            Double.parseDouble(editedExpenditure.getMoney().toString()));
                 } else if ((month2 == month) && (day2 >= day)) {
                     total = total + Float.parseFloat(editedExpenditure.getMoney().toString());
                     dpd.setValue(editedExpenditure.getDescription().toString() + ":"
-                                    + Integer.parseInt(editedExpenditure.getMoney().toString()),
-                            Integer.parseInt(editedExpenditure.getMoney().toString()));
+                                    + Double.parseDouble(editedExpenditure.getMoney().toString()),
+                            Double.parseDouble(editedExpenditure.getMoney().toString()));
                 }
             }
             index++;
